msm: qdss: add new external qdss header file
This will serve to provide qdss driver apis that expose the qdss
driver services to the qdss clients. Qdss clients are expected to
include this header file.

This header file will also provide access to apis implemented
in jtag.c that are responsible for debug and etm register save and
restore.

Change-Id: I9ee031d7862dcdb6b81ae37e88b92a8b0db01b20
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>