gpio: qpnp-gpio: Remove config param inv_int_pol
This configuration option should never have existed. It does not
exist in the qpnp-gpio block. It was mistakenly being used to
represent the 'output_value' parameter within the code. Fix such
occurrences so that the semantics are consistent.

Also move qpnp-gpio documentation to a more meaningful location.
It belongs in the gpio directory and not spmi.

Change-Id: I48b37aef02793934f60b9886347068cf37331c3f
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>