ASoC: msm: Add supplementary service for VoLTE and voice call.
Add support for call hold and resume for Voice and VoLTE calls.
This is added to support new events from user space for voice
driver. These new events are used to switch between Voice and
VoLTE calls. These events are handled in trigger function of
driver which puts vocproc to standby mode for pause and starts
when resume is passed to trigger function of driver.

Change-Id: Id8774515fdb8437846e7a8864002f4be800e66e1
Signed-off-by: Venkat Sudhir <vsudhir@codeaurora.org>