scheduler: compute time-average nr_running per run-queue
Compute the time-average number of running tasks per run-queue for a
trailing window of a fixed time period. The detla add/sub to the
average value is weighted by the amount of time per nr_running value
relative to the total measurement period.

This also fixes the Intellidemand governor eorrors as it uses avg_nr_running

Conflicts:
	arch/arm/configs/mapphone_OCE_defconfig
	include/linux/sched.h
	kernel/sched.c
	kernel/sched_debug.c